refactor(Game): split components
- move components into separate files (separation of concerns)
