+++ src/TicTacToe/Square.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Square extends React.Component {
+  render() {
+    return <button className="square">{/* TODO */}</button>;
+  }
+}
+
+export default Square;
